move classes to a named package, fix the class not found in jar, and add a more detailed gradle jar build phase
