+++ Dockerfile
@@ -10,3 +10,4 @@ RUN apt-get update && apt-get install -y \
 
 WORKDIR /root
 RUN ["/bin/bash", "-c", "mkdir data && cd data && while read i; do git clone $i; done < <(curl -s https://api.github.com/orgs/datasets/repos?per_page=100 | jq -r '.[].clone_url')" ]
+RUN ["/bin/bash", "-c", "git clone https://github.com/rlperez/screen.git && cd screen && gradle "]
