+++ build.gradle
@@ -17,6 +17,16 @@ apply plugin: 'kotlin'
 
 sourceCompatibility = 1.8
 
+jar {
+    manifest {
+        attributes "Main-Class": "com.rlperez.ScreenApplication"
+    }
+
+    from {
+        configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
+    }
+}
+
 repositories {
     mavenCentral()
 }
