+++ src/main/kotlin/com/rlperez/Solution3.kt
@@ -1,3 +1,5 @@
+package com.rlperez
+
 import java.io.File
 
 class Solution3 {
